chore: add classes and packages

+++ ToDoMicroService/src/main/java/com/todomicroservice/repository/ToDoRepository.java
@@ -1,6 +1,7 @@
 package com.todomicroservice.repository;
 
 import com.todomicroservice.entity.ToDo;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ToDoRepository extends JpaRepository<ToDo, Integer> {
+public interface ToDoRepository extends JpaRepository<ToDo, Long> {
 }
